Change deletion on Windows machine

+++ src/main/java/io/openmanufacturing/ame/repository/strategy/LocalFolderResolverStrategy.java
@@ -29,6 +29,7 @@ import java.util.stream.Stream;
 
 import javax.annotation.Nonnull;
 
+import org.apache.commons.exec.OS;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -115,8 +116,8 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
       }
 
       final AspectModelUrn aspectModelUrn = AspectModelUrn.fromUrn( urn );
-      return aspectModelUrn.getNamespace() + File.separator + aspectModelUrn.getVersion() + File.separator +
-            aspectModelUrn.getName() + applicationSettings.getFileType();
+      return aspectModelUrn.getNamespace() + File.separator + aspectModelUrn.getVersion() + File.separator
+            + aspectModelUrn.getName() + applicationSettings.getFileType();
    }
 
    @Override
@@ -192,10 +193,11 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
 
       final List<String> endFilePaths = getEndFilePaths( rootSharedFolder, file );
 
-      final Map<String, List<String>> namespacePathMapping = endFilePaths
-            .stream()
-            .map( LocalFolderResolverStrategy::transformToValidModelDirectory )
-            .collect( Collectors.groupingBy( this::extractNamespaceAndVersion, toList() ) );
+      final Map<String, List<String>> namespacePathMapping = endFilePaths.stream()
+                                                                         .map( LocalFolderResolverStrategy::transformToValidModelDirectory )
+                                                                         .collect( Collectors.groupingBy(
+                                                                               this::extractNamespaceAndVersion,
+                                                                               toList() ) );
 
       retainOnlyTurtleFileName( namespacePathMapping );
 
@@ -208,10 +210,8 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
    private List<String> getEndFilePaths( @Nonnull final String rootSharedFolder, @Nonnull final File file ) {
       try ( final Stream<Path> paths = getAllSubFilePaths( file.toPath() ) ) {
 
-         return paths.filter( this::isPathRelevant )
-                     .map( Path::toString )
-                     .map( path -> excludeStandaloneFiles( rootSharedFolder, path ) )
-                     .filter( StringUtils::isNotBlank )
+         return paths.filter( this::isPathRelevant ).map( Path::toString )
+                     .map( path -> excludeStandaloneFiles( rootSharedFolder, path ) ).filter( StringUtils::isNotBlank )
                      .toList();
       } catch ( final IOException e ) {
          throw new FileReadException( "Can not read shared folder file structure", e );
@@ -227,11 +227,8 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
          return paths.filter( path -> {
                         final File parentDir = getFileInstance( path.toString() );
                         return !parentDir.isDirectory() && !path.toString().endsWith( ModelUtils.TTL_EXTENSION );
-                     } )
-                     .map( Path::toString )
-                     .map( path -> excludeStandaloneFiles( rootSharedFolder, path ) )
-                     .filter( StringUtils::isNotBlank )
-                     .toList();
+                     } ).map( Path::toString ).map( path -> excludeStandaloneFiles( rootSharedFolder, path ) )
+                     .filter( StringUtils::isNotBlank ).toList();
       } catch ( final IOException e ) {
          throw new FileReadException( "Can not read shared folder file structure", e );
       }
@@ -257,14 +254,11 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
    }
 
    private List<ValidFile> getListOfLocalPackageInformation( final List<String> filePath, final String storagePath ) {
-      return filePath
-            .stream()
-            .map( path -> {
-               final String aspectModelFile = transformToValidModelDirectory( path );
-               final String aspectModel = getModelAsString( aspectModelFile, storagePath );
-               return new ValidFile( aspectModelFile, aspectModel );
-            } )
-            .toList();
+      return filePath.stream().map( path -> {
+         final String aspectModelFile = transformToValidModelDirectory( path );
+         final String aspectModel = getModelAsString( aspectModelFile, storagePath );
+         return new ValidFile( aspectModelFile, aspectModel );
+      } ).toList();
    }
 
    /**
@@ -289,13 +283,11 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
     */
    private void retainOnlyTurtleFileName( @Nonnull final Map<String, List<String>> pathTurtleFilesMap ) {
       for ( final Map.Entry<String, List<String>> entry : pathTurtleFilesMap.entrySet() ) {
-         final List<String> collect = entry.getValue().stream()
-                                           .filter( value -> !value.equals( entry.getKey() ) )
+         final List<String> collect = entry.getValue().stream().filter( value -> !value.equals( entry.getKey() ) )
                                            .map( value -> value.replaceAll( entry.getKey(), StringUtils.EMPTY ) )
                                            .map( value -> value.replace(
                                                  LocalFolderResolverUtils.NAMESPACE_VERSION_NAME_SEPARATOR,
-                                                 StringUtils.EMPTY ) )
-                                           .toList();
+                                                 StringUtils.EMPTY ) ).toList();
 
          entry.setValue( collect );
       }
@@ -374,8 +366,7 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
     * @param storagePath - path to storage files.
     * @return Aspect Model urn
     */
-   protected AspectModelUrn getAspectModelUrn( @Nonnull final String turtleData,
-         final @Nonnull String storagePath ) {
+   protected AspectModelUrn getAspectModelUrn( @Nonnull final String turtleData, final @Nonnull String storagePath ) {
       return inMemoryStrategy( turtleData, ValidationProcess.getEnum( storagePath ) ).getAspectModelUrn();
    }
 
@@ -390,8 +381,8 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
          final @Nonnull String storagePath ) {
       final AspectModelUrn aspectModelUrn = getAspectModelUrn( turtleData, storagePath );
 
-      return aspectModelUrn.getNamespace() + File.separator + aspectModelUrn.getVersion() + File.separator +
-            aspectModelUrn.getName();
+      return aspectModelUrn.getNamespace() + File.separator + aspectModelUrn.getVersion() + File.separator
+            + aspectModelUrn.getName();
    }
 
    /**
@@ -433,8 +424,8 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
       try {
          return new String( Files.readAllBytes( storeFile.toPath() ) );
       } catch ( final IOException e ) {
-         throw new FileNotFoundException( String.format( "Cannot read file at the following path: %s",
-               storeFile.toPath() ), e );
+         throw new FileNotFoundException(
+               String.format( "Cannot read file at the following path: %s", storeFile.toPath() ), e );
       }
    }
 
@@ -446,10 +437,13 @@ public class LocalFolderResolverStrategy implements ModelResolverStrategy {
     */
    private void deleteFile( @Nonnull final File file ) {
       try {
-         FileUtils.forceDelete( file );
+         if ( OS.isFamilyWindows() ) {
+            FileUtils.forceDeleteOnExit( file );
+         } else {
+            FileUtils.forceDelete( file );
+         }
       } catch ( final IOException e ) {
-         throw new FileNotFoundException( String.format( "File %s was not deleted successfully.", file.toPath() ),
-               e );
+         throw new FileNotFoundException( String.format( "File %s was not deleted successfully.", file.toPath() ), e );
       }
    }
